Check for ConnectionRefusedError while working with RPCs
RPCs here are synchronous and assume connections are established.
However, things on workstation tend to fail during demos.

Hence this is a fallback in case RPCs are made to an already down
endpoint.

The error message is kept as generic as possible which does not
let the client know about the location distribution which is inherent
a distributed file system.